Create failedP5.html
P5 was chosen for the intial prototype because of the familiarity and ease of learning with the interface and processing as a language. 

Early Tests Done with P5. 
Errors were encountered with the loading of an .obj file (an early stage feature) and there was not very sufficient online support. 
The lack of online knowledge and robustness related to p5 and 3D lead me to switch to Three.js, which also has more robust features.